refactor: Delete user role in SupabaseUserData.js
Delete the user role by adding a new function `deleteUserRole` in `SupabaseUserData.js`. This function uses the Supabase client to delete the user role from the `user_roles` table based on the provided `userId`. If an error occurs during the deletion, it will be thrown. The function returns the deleted data.

Also, add the `UserDeleteRoleSuccessNotification` component to the list of exported components in `Notifications.jsx`.

This commit refactors the code to improve code readability and consistency in the `SupabaseUserData.js` file.