refactor: Update stacked.jsx to use next/image for profile picture
This commit updates the stacked.jsx file to use the next/image component for displaying the profile picture. The previous implementation used the img tag, but now it has been replaced with the Image component from the next/image package. This change improves the performance and optimization of image loading in the application.