refactor: Add getInvoicePaidStatus and updateInvoicePaidStatus functions
This commit adds two new functions, getInvoicePaidStatus and updateInvoicePaidStatus, to the SupabaseUserData module. These functions are responsible for retrieving and updating the paid status of an invoice in the Supabase database. They enhance the functionality of the application by providing the ability to track and manage the payment status of invoices.