refactor: Add authorizer name and user ID to updateInvoicePaidStatus function
This commit adds the authorizer name and user ID parameters to the updateInvoicePaidStatus function in the SupabaseUserData module. It allows for updating the paid status and status of an invoice, as well as specifying the authorizer's name and user ID. This enhancement improves the functionality of the application by providing more comprehensive information about the invoice authorization process.